Get all flags of DART for conversion
