+++ src/Magnum/DartIntegration/CMakeLists.txt
@@ -35,12 +35,10 @@ configure_file(${CMAKE_CURRENT_SOURCE_DIR}/configure.h.cmake
                ${CMAKE_CURRENT_BINARY_DIR}/configure.h)
 
 set(MagnumDartIntegration_SRCS
-    ConvertShapeNode.cpp
     Object.cpp
     World.cpp)
 
 set(MagnumDartIntegration_HEADERS
-    ConvertShapeNode.h
     DartIntegration.h
     Object.h
     World.h
